# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quantum FX Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset link sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Dashboard
...
  /bots/{id}:
    get:
      operationId: getBot
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bots/{id}/toggle:
    post:
      operationId: toggleBot
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
  /bots/{id}/analytics/{period}:
    get:
      operationId: getBotAnalytics
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: period
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot analytics chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartPoint"
  /marketplace/bots:
    get:
      operationId: listMarketplaceBots
      tags: [bots]
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Marketplace bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceBot"
  /marketplace/bots/{id}/purchase:
    post:
      operationId: purchaseBot
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Cashier
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Trade
...
    RegisterInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        referralCode:
          type: ["string", "null"]
      required: [fullName, email, password]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
      required: [email]
    ResetPasswordInput:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required: [token, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        kycStatus:
          type: string
        createdAt:
          type: string
      required: [id, fullName, email, kycStatus, createdAt]

    # Dashboard
...
    ChartPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
      required: [date, value]
...
    Bot:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
        profitToday:
          type: number
        winRate:
          type: number
        totalTrades:
          type: integer
        iconUrl:
          type: ["string", "null"]
        category:
          type: string
      required: [id, name, status, profitToday, winRate, totalTrades, category]
    BotDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
        profitToday:
          type: number
        profitTotal:
          type: number
        winRate:
          type: number
        totalTrades:
          type: integer
        iconUrl:
          type: ["string", "null"]
        category:
          type: string
        description:
          type: string
        performance:
          type: number
      required: [id, name, status, profitToday, profitTotal, winRate, totalTrades, category, description, performance]
    MarketplaceBot:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        winRate:
          type: number
        category:
          type: string
        riskLevel:
          type: string
        description:
          type: string
        iconUrl:
          type: ["string", "null"]
        isPurchased:
          type: boolean
      required: [id, name, price, winRate, category, riskLevel, description, isPurchased]

    # Cashier
...