# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quantum FX Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  kycStatus: string;
  createdAt: string;
}